add text-hiding-function
